requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pydantic>=1.8.2
pydantic-settings>=2.0.0
langchain>=0.1.0
langgraph>=0.0.10
pinecone-client>=2.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
pytest>=6.2.5
httpx>=0.23.0
docker>=5.0.3 